ENH: datetime.time converters for plotting
