+++ pandas/src/inference.pyx
@@ -103,6 +103,9 @@ cdef inline bint is_datetime(object o):
 cdef inline bint is_date(object o):
     return PyDate_Check(o)
 
+cdef inline bint is_time(object o):
+    return PyTime_Check(o)
+
 def is_bool_array(ndarray values):
     cdef:
         Py_ssize_t i, n = len(values)
@@ -240,6 +243,14 @@ def is_date_array(ndarray[object] values):
             return False
     return True
 
+def is_time_array(ndarray[object] values):
+    cdef int i, n = len(values)
+    if n == 0:
+        return False
+    for i in range(n):
+        if not is_time(values[i]):
+            return False
+    return True
 
 def maybe_convert_numeric(ndarray[object] values, set na_values,
                           convert_empty=True):
